requirements.txt
langchain==1.0.2
langgraph==1.0.2
langchain-openai==1.0.1
openai==2.6.1
chromadb==1.3.0
requests==2.32.5
pandas==2.3.3
fastapi==0.120.2
uvicorn==0.38.0
python-dotenv==1.2.1
cachetools==6.2.1
pydantic==2.12.3
typing-extensions==4.15.0
aiohttp==3.11.10
